refactor: extract useDayjs
