+++ src/App.tsx
@@ -1,31 +1,9 @@
-import { useEffect, useState } from "react"
-import dayjs, { Dayjs } from "dayjs"
-import { useInterval } from "@mantine/hooks"
 import { Helmet, HelmetProvider } from "react-helmet-async"
 
 import { isoDate } from "./isoDate.ts"
 import { clockHours } from "./ClockIcon.tsx"
+import { useDayjs } from "./useDayjs.tsx"
 import "./App.css"
-
-/**
- * returns `new Date()` every interval
- *
- * @param interval interval in ms
- * @returns date object
- */
-export const useDate = (interval: number = 1000): Date => {
-  const [date, setDate] = useState(new Date())
-  const { start } = useInterval(() => setDate(new Date()), interval)
-
-  useEffect(() => start())
-
-  return date
-}
-
-export const useDayjs = (interval: number = 1000): Dayjs => {
-  const date = useDate(interval)
-  return dayjs(date)
-}
 
 export const App = () => {
   const dayjs = useDayjs()
